refactor: Image 도메인의 패키지 구조 및 서비스 로직 계층 분리
- ImageService 하위 계층으로 ImageUploader, ImageReader, ImageDeleter를 구성한다.
- image 패키지를 application, domain, infra로 구분한다.
