+++ src/main/java/com/ssh/dartserver/domain/image/infra/ImageRepository.java
@@ -1,5 +1,6 @@
-package com.ssh.dartserver.domain.image;
+package com.ssh.dartserver.domain.image.infra;
 
+import com.ssh.dartserver.domain.image.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
